klusters: accumulate the GPU aggregation in float (drop the throttled F2D)
The FP32 change dropped the Mahalanobis kernel ~1591 -> ~985 ms as intended, but
the aggregate phase rose ~65 -> ~90 ms even though d_prob shrank to FP32 — the
opposite of the expected memory-bound speedup. Cause: the reduction promoted each
posterior to double (sum += (double)prob[...]) across nSpikes x nClusters ~1.86e9
reads, and that float->double conversion runs through the throttled FP64 path on
the GB202 die, so the phase became conversion-bound rather than bandwidth-bound.

Accumulate in float instead. The posteriors are already FP32, so double
accumulation bought no real accuracy (the ~5e-4 FP32 posterior error dominates),
while the mean's float rounding error is ~1e-5 relative — well under that floor.
errOut stays double (a single float->double per matrix cell, ~nClusters^2, is
negligible). This removes the per-read conversion and lets the aggregate run at
FP32 bandwidth. Corrects the earlier decision to keep the accumulation in double.

Expect aggregate ~90 -> ~35 ms (FULL ~1.2 -> ~1.15 s), with no visible change to
the qualitative matrix.

+++ src/klusters/src/groupingassistant_cuda.cu
@@ -216,12 +216,17 @@ __global__ void cuda_aggregate_kernel(
     const int  n  = nb[i];
     for (int j = threadIdx.x; j < nbClusters; j += blockDim.x) {
         if (ig || ignoreFlags[j] || n <= 0) { errOut[(size_t)i * nbClusters + j] = 0.0; continue; }
-        double sum = 0.0;
+        // Accumulate in float: the posteriors are already FP32, so promoting each
+        // read to double buys no accuracy (the ~5e-4 FP32 floor dominates) while
+        // costing a per-element float->double conversion x nSpikes x nClusters,
+        // which on the throttled-FP64 GB202 runs through the scarce FP64 path.
+        // The mean's float rounding error is ~1e-5 relative, well under that floor.
+        float sum = 0.0f;
         for (int s = 0; s < n; ++s) {
             int row = featRow[f + s];
-            sum += (double)prob[(size_t)row * nbClusters + j];
+            sum += prob[(size_t)row * nbClusters + j];
         }
-        errOut[(size_t)i * nbClusters + j] = sum / (double)n;
+        errOut[(size_t)i * nbClusters + j] = (double)(sum / (float)n);
     }
 }
